refactor: remove NFTCardStyledAssetPlayer
