+++ packages/shared/src/UI/components/index.ts
@@ -33,7 +33,6 @@ export * from './MiniNetworkSelector/index.js'
 export * from './NetworkIcon/index.js'
 export * from './NetworkTab/index.js'
 export * from './NFTCard/index.js'
-export * from './NFTCardStyledAssetPlayer/index.js'
 export * from './NFTList/index.js'
 export * from './PersonaAction/index.js'
 export * from './PluginCardFrame/index.js'
